+++ zenoh/src/api/session.rs
@@ -88,7 +88,6 @@ use super::{
     publisher::Publisher,
     publisher::{MatchingListenerState, MatchingStatus},
     sample::SourceInfo,
-    selector::TIME_RANGE_KEY,
 };
 use crate::{
     api::selector::PredefinedParameters,
